+++ src/main/java/com/codingbee/snn4j/interface_implementations/random_weight_generators/XavierUniformGen.java
@@ -13,21 +13,21 @@ import java.util.Random;
 public class XavierUniformGen implements RandomWeightGenerator {
     Random gen = new Random();
     @Override
-    public double getWeight(int inputs, int outputs) throws IncorrectDataException {
+    public float getWeight(int inputs, int outputs) throws IncorrectDataException {
         if (inputs < 1 || outputs < 1){
             throw new IncorrectDataException("The layer sizes, the weights are between, have to be larger than 0");
         }
         double limit = Math.sqrt(6.0) / Math.sqrt(inputs + outputs);
-        return gen.nextDouble() * 2 * limit - limit;
+        return (float) (gen.nextDouble() * 2 * limit - limit);
     }
 
     @Override
-    public double getHiddenLayerBias() {
+    public float getHiddenLayerBias() {
         return 0;
     }
 
     @Override
-    public double getOutputLayerBias() {
+    public float getOutputLayerBias() {
         return 0;
     }
 }
