+++ src/main/java/com/codingbee/snn4j/interfaces/RandomWeightGenerator.java
@@ -3,7 +3,7 @@ package com.codingbee.snn4j.interfaces;
 import com.codingbee.snn4j.exceptions.IncorrectDataException;
 
 public interface RandomWeightGenerator {
-    double getWeight(int inputs, int outputs) throws IncorrectDataException;
-    double getHiddenLayerBias();
-    double getOutputLayerBias();
+    float getWeight(int inputs, int outputs) throws IncorrectDataException;
+    float getHiddenLayerBias();
+    float getOutputLayerBias();
 }
